펫추가 품종 UI 진행중
