+++ src/components/Input/Input.tsx
@@ -1,14 +1,17 @@
 import React from 'react';
 import { Pressable, TextInput, View } from 'react-native';
 import { useForm, Controller } from 'react-hook-form'
+import { Colors } from '../../styles/Colors';
 
 interface IProps {
+    name: string;
+    defaultValue?: string;
     onPress?: () => void,
     placeholder: string;
 }
 
 const Input = (props: IProps) => {
-    const { onPress, placeholder } = props;
+    const { name, defaultValue, onPress, placeholder } = props;
 
     const { control, handleSubmit, formState: { errors } } = useForm();
 
@@ -23,10 +26,14 @@ const Input = (props: IProps) => {
                         onBlur={onBlur}
                         onChangeText={onChange}
                         value={value}
+                        style={{
+                            borderWidth: 1,
+                        }}
                     />
                     </Pressable>
                 )}
-                name="lastName"
+                name={name}
+                defaultValue={defaultValue || ''}
             />
         </View>
     );
