--- src/screens/myPage/petManagement/AddAnotherAnimal.tsx
@@ -1,20 +1,0 @@
-import React from 'react';
-import { View, Image, StyleSheet } from 'react-native';
-import asset from "../../../assets/png";
-
-const AddAnotherAnimal = () => {
-    return (
-        <View style={styles.Container}>
-            <Image source={asset.profile} />
-        </View>
-    );
-};
-
-export default AddAnotherAnimal;
-
-const styles = StyleSheet.create({
-    Container: {
-        alignItems: "center",
-        marginTop: 60,
-    },
-});
