+++ types/RootStackParams.ts
@@ -10,5 +10,6 @@ export type RootStackParams = {
   Notice: undefined;
   PatManagement: undefined;
   Profile: undefined;
-  AddPet: undefined;
+  AddPetScreen: undefined;
+  BreedType: undefined;
 };
